Refactor `Command`, a bit more cleanup (#97)
* Simplify command, remove `CommandCode`

* Remove `Debug` bound

* Reduce indirection a bit

* Combine impl blocks

* Use full paths for core::fmt

Makes for easier enable/disable/whatever in the future

+++ src/pdu_loop/frame_element/sendable_frame.rs
@@ -88,7 +88,7 @@ impl<'sto> SendableFrame<'sto> {
 
         let buf = le_u16(header.0, buf);
 
-        let buf = le_u8(frame.command.code() as u8, buf);
+        let buf = le_u8(frame.command.code(), buf);
         let buf = le_u8(frame.index, buf);
 
         // Write address and register data
